Fix UI dialogs, enums, MensajeReceptor, migrations, and docs
Tipo_CodigoImpuesto and Tipo_TarifaIVA enum updates for v4.4 tariff codes. New MensajeReceptorService for electronic payment receipt messaging. ExchangeRatesController update. PDFGenerator and ProgramadorTareas fixes. XmlEncabezadoFlattener update. CreateArticuloDialog and EditArticuloDialog UI improvements for exoneracion fields. Recibos page styling fixes. DB migrations for cabys impuesto string type and TipoCambio BigDecimal. application.properties update. Added FUTURE_ENHANCEMENTS.md documenting latent gaps (esServicio, ISC, tariff mappings).

+++ mercurius-quarkus/src/main/java/Controllers/Api/Accounting/ExchangeRatesController.java
@@ -8,6 +8,7 @@ import jakarta.inject.Inject;
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -72,7 +73,7 @@ public class ExchangeRatesController {
     public static class ExchangeRateDTO {
         public Long id;
         public LocalDateTime fecha;
-        public double valorCompra;
-        public double valorVenta;
+        public BigDecimal valorCompra;
+        public BigDecimal valorVenta;
     }
 }
